refactor(payment): Payment 관련 클래스 및 테스트 패키지 구조 변경
Payment 관련 클래스와 테스트를 common 패키지에서 payment 패키지로 이동하여 모듈화 및 패키지 구조 정리.
