+++ src/test/java/roomescape/reservation/ReservationServiceTest.java
@@ -24,9 +24,9 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.dto.LoginMember;
-import roomescape.common.PaymentManager;
+import roomescape.payment.PaymentManager;
 import roomescape.common.TimeManager;
-import roomescape.common.dto.PaymentRequest;
+import roomescape.payment.PaymentRequest;
 import roomescape.exception.custom.reason.payment.PaymentConfirmException;
 import roomescape.exception.custom.reason.reservation.ReservationConflictException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsMemberException;
